Combine libraries
